This demo shows how to inject a dependency through the component.
This is a great way to work with unit testing. It is also a nice way to bootstrap injection at runtime. For example what if Gradle generates a json config at runtime, then application can obtain an object from it, and pass it directly to the AppComponent. (More to come!)